gcp-data-dropzone-challenge/cloud-function/transform-json/requirements.txt
google-cloud-storage
SQLAlchemy==1.3.12
PyMySQL==0.9.3
